add helper function for sending new email
