+++ webapp/.env.development
@@ -1,6 +1,7 @@
 # https://vite.dev/guide/env-and-mode.html#env-files
 DATABASE_URL=${DATABASE_URL:-postgres://${POSTGRES_USER:-astralbeam}:${POSTGRES_PASSWORD:-astralbeam}@${POSTGRES_HOST:-127.0.0.1}:${POSTGRES_HOST_PORT:-5432}/${POSTGRES_DB:-astralbeam}}
 VALKEY_URL=${VALKEY_URL:-valkey://${VALKEY_HOST:-127.0.0.1}:${VALKEY_HOST_PORT:-6379}}
+APP_BASE_URL=${APP_BASE_URL:-http://localhost:3000}
 EMAIL_PROVIDER=${EMAIL_PROVIDER:-resend}
 # Resend's shared sandbox sender, which needs no verified domain. https://resend.com/docs/dashboard/emails/send-test-emails
 EMAIL_FROM_ADDRESS=${EMAIL_FROM_ADDRESS:-onboarding@resend.dev}
