+++ webapp/.env.example
@@ -1,6 +1,7 @@
 # Local service defaults come from .env.development. Copy the variables you need to override into .env.local or provide them through the shell or deployment environment.
 # DATABASE_URL=
 # VALKEY_URL=
+# APP_BASE_URL=
 
 # Required by /api/chat; set it in .env or the shell, never in a committed file.
 # OPENAI_API_KEY=
