+++ webapp/src/lib/config.ts
@@ -1,0 +1,1 @@
+export const APP_NAME = "AstralBeam"
